# schemas/openapi/platform.openapi.yaml
openapi: 3.1.0
info:
  title: Modpack Compatibility Platform API
  version: 0.2.0
  description: Phase 0 implementation-grade contract for the documented public API subset.
servers:
  - url: http://localhost:8080
    description: Local development gateway
paths:
...
  /v1/analyses/{analysisId}:
    get:
      summary: Get analysis summary
      operationId: getAnalysis
      parameters: [{ $ref: '#/components/parameters/AnalysisId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnalysisSummary' }
        default: { $ref: '#/components/responses/Error' }
...
components:
  parameters:
...
    AnalysisId: { in: path, name: analysisId, required: true, schema: { type: string } }
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            trace_id: { type: string }
...
    AnalysisSummary:
      type: object
      required: [analysis_id, project_id, workspace_id, status, trigger, analysis_mode, created_at]
      properties:
        analysis_id: { type: string }
        project_id: { type: string }
        workspace_id: { type: string }
        organization_id: { type: string }
        pack_snapshot_id: { type: string }
        status: { type: string, enum: [queued, running, completed, failed, cancelled] }
        trigger: { type: string, enum: [manual, webhook, scheduled, api] }
        analysis_mode: { type: string, enum: [fast, standard, deep] }
        score: { type: number, format: float }
        verdict: { type: string, enum: [known_incompatible, likely_incompatible, mixed_or_conditional, no_known_issue_found, insufficient_evidence] }
        confidence:
          type: object
          required: [score, band, explanation, primary_drivers]
          properties:
            score: { type: number, minimum: 0, maximum: 1 }
            band: { type: string, enum: [very_high, high, medium, low, very_low] }
            explanation: { type: string }
            primary_drivers: { type: array, items: { type: string } }
        coverage_status: { type: string, enum: [high_confidence_coverage, best_effort_coverage, outside_validated_coverage] }
        serving_mode: { type: string, enum: [snapshot_exact_pack, snapshot_retrieval_assembly, legacy_live_analysis, offline_benchmark] }
        knowledge_snapshot_version: { type: string }
        explanation: { type: string }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        counts:
          type: object
          properties:
            critical: { type: integer }
            high: { type: integer }
            medium: { type: integer }
            low: { type: integer }
            info: { type: integer }
...